requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic>=2.9.0
python-dotenv==1.0.1
requests==2.31.0
openai>=1.99.5
python-multipart==0.0.9
python-dateutil==2.9.0